adapt MainModal to accept component and delete redundant CreateNote  gql queries
