+++ libs/expo/betterangels/src/lib/screens/Client/Client.graphql
@@ -156,33 +156,3 @@ mutation DeleteClientDocument($id: ID!) {
     }
   }
 }
-
-mutation CreateNote($data: CreateNoteInput!) {
-  createNote(data: $data) {
-    ... on OperationInfo {
-      messages {
-        kind
-        field
-        message
-      }
-    }
-
-    ... on NoteType {
-      id
-      purpose
-      publicDetails
-      clientProfile {
-        id
-        firstName
-        lastName
-        email
-      }
-      createdAt
-      createdBy {
-        id
-        username
-        email
-      }
-    }
-  }
-}
