+++ libs/expo/betterangels/src/lib/ui-components/ClientCardModal.tsx
@@ -1,9 +1,8 @@
 import { FilePlusIcon, UploadIcon } from '@monorepo/expo/shared/icons';
 import { useRouter } from 'expo-router';
-import { useSnackbar } from '../hooks';
 import { ClientProfilesQuery } from '../screens/Clients/__generated__/Clients.generated';
-import { useCreateNoteMutation } from '../screens/Home/__generated__/ActiveClients.generated';
-import MainModal from './MainModal';
+import { CreateClientInteractionBtn } from './CreateClientInteraction';
+import { MainModal, MainModalActionBtnBody } from './MainModal';
 
 interface IMainPlusModalProps {
   closeModal: () => void;
@@ -13,49 +12,30 @@ interface IMainPlusModalProps {
 
 export default function ClientCardModal(props: IMainPlusModalProps) {
   const { isModalVisible, closeModal, clientProfile } = props;
-  const [createNote] = useCreateNoteMutation();
   const router = useRouter();
-  const { showSnackbar } = useSnackbar();
 
-  async function createNoteFunction(id: string) {
-    try {
-      const { data } = await createNote({
-        variables: {
-          data: {
-            clientProfile: id,
-          },
-        },
-      });
-      if (data?.createNote && 'id' in data.createNote) {
-        router.navigate(`/add-note/${data.createNote.id}`);
+  const CreateInteractionBtn = (
+    <CreateClientInteractionBtn
+      clientProfileId={clientProfile.id}
+      onCreated={(noteId) => {
         closeModal();
-      }
-    } catch (err) {
-      console.error(err);
-
-      showSnackbar({
-        message: `Sorry, there was an error creating a new interaction.`,
-        type: 'error',
-      });
-    }
-  }
+        router.navigate(`/add-note/${noteId}`);
+      }}
+      style={{ width: '100%' }}
+    >
+      <MainModalActionBtnBody title="Add Interaction" Icon={FilePlusIcon} />
+    </CreateClientInteractionBtn>
+  );
 
   const ACTIONS = [
-    {
-      title: 'Add Interaction x',
-      Icon: FilePlusIcon,
-      onPress: () => {
-        if (clientProfile) {
-          createNoteFunction(clientProfile.id);
-        }
-      },
-    },
+    CreateInteractionBtn,
     {
       title: 'Upload Documents',
       Icon: UploadIcon,
       route: `/client/${clientProfile.id}?newTab=Docs`,
     },
   ];
+
   return (
     <MainModal
       closeButton
