+++ libs/expo/betterangels/src/lib/ui-components/MainModal/index.ts
@@ -1,0 +1,4 @@
+export { MainModal } from './MainModal';
+export type { TMainModalAction } from './MainModal';
+export { MainModalActionBtn } from './MainModalActionBtn';
+export { MainModalActionBtnBody } from './MainModalActionBtnBody';
