+++ libs/expo/betterangels/src/lib/ui-components/index.ts
@@ -18,6 +18,7 @@ export { FileThumbnail } from './FileThumbnail/FileThumbnail';
 export { default as Header } from './Header';
 export { default as LoginForm } from './LoginForm';
 export { default as MainContainer } from './MainContainer';
+export * from './MainModal';
 export { default as MainPlusModal } from './MainPlusModal';
 export { default as MainScrollContainer } from './MainScrollContainer';
 export { default as Modal } from './Modal';
